+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
-- `org`: `Person` new `key` and `identities` properties.
+- `org`: `Person` new `key`, `addresses`, and `identities` properties.
 
 ## [v0.218.0] - 2025-06-12